+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # v2.0.0
 
 - Updates @vcmap/core and @vcmap/ui to verison 6.0
+- Fixes 3D Height sketch theme to match dark mode
 
 # v1.0.8